License fix: Analysis results to common for MIT licensing
They are re-exported from their previous location since they are mostly
used from GPL code. Only the MIT import is changed to the new location.
